Added default attributes
These could be put into a Common or Attributes namespace if they pollute
Thoth too much.
Also good to note is some attributes are only allowed to be added to
certain elements. Should this be taken care of on the attribute side,
element side, or left alone so users can shoot their own foot.